move utils function to indexer utils due to import issues
